+++ xenian_bot/commands/base.py
@@ -71,29 +71,19 @@ class BaseCommand:
             command = {
                 'title': command.get('title', None) or command['command'].__name__.capitalize().replace('_', ' '),
                 'description': command.get('description', ''),
-                'command_name': command.get('command_name', None) or command['command'].__name__,
+                'command_name': command.get('command_name', command['command'].__name__),
                 'command': command['command'],
-                'handler': command.get('handler', None) or CommandHandler,
-                'options': command.get('options', None),
+                'handler': command.get('handler', CommandHandler),
+                'options': command.get('options', {}),
                 'hidden': command.get('hidden', False),
                 'args': command.get('args', None)
             }
-            # Set options if not yet set
-            if command['options'] is None:
-                command['options'] = {'callback': command['command'], 'command': command['command_name']}
 
-            # Set CommandHandler options if not yet set
-            if command['handler'] == CommandHandler and command['options'].get('callback', None) is None:
-                command['options']['callback'] = command['command']
             if command['handler'] == CommandHandler and command['options'].get('command', None) is None:
                 command['options']['command'] = command['command_name']
 
-            # Set MessageHandler options if not yet set
-            if command['handler'] == MessageHandler and command['options'].get('callback', None) is None:
-                command['options']['callback'] = command['command']
-
             # Set CallbackQueryHandler options if not yet set
-            if command['handler'] == CallbackQueryHandler and command['options'].get('callback', None) is None:
+            if command['options'].get('callback', None) is None:
                 command['options']['callback'] = command['command']
 
             updated_commands.append(command)